Add option to force GvhProjectSettings.xml line endings
With different team members working with Unity on Windows and on Mac,
the line endings of this file keep changing, causing spurious changes
to show up in version control. This option forces the line endings to
a configured value, but uses the platform default if the option is
not supplied.